+++ changelog.md
@@ -1,5 +1,6 @@
 ## 1.7.1
   - Add `Display name` to customer pop-up
+  - Add button to check geolocation issue for job requests and estimates list
 
 ## 1.7.0
   - Use pop-up box for create/edit customer in request and estimate forms